Ensure that we always serverlatest theme css
We now call setCacheWithRevalidationHeaders when serving theme css to make sure that the browser doesn't use an older copy of the theme (now at a minimum the browser must get back a 304 to use a cached copy).

There are several other places in the codebase where a similar change is likely warranted.